Integrated razorpay in payment service and started with SAGA pattern

+++ cart/src/main/java/com/ecommerce/cart/dto/CartItemsResponse.java
@@ -8,6 +8,6 @@ import java.math.BigDecimal;
 public class CartItemsResponse {
     private String productId;
     private Integer quantity;
-    private BigDecimal price;
-    private BigDecimal totalPrice;
+    private Long price;
+    private Long totalPrice;
 }
