+++ cart/src/main/java/com/ecommerce/cart/dto/CartResponse.java
@@ -11,5 +11,5 @@ import java.util.List;
 public class CartResponse {
     private String userId;
     private List<CartItemsResponse> items;
-    private BigDecimal cartTotal;
+    private Long cartTotal;
 }
